Open plans directly when computer features need payment
Paywalled computer flows now open the plan picker with one-click
checkout instead of a dead-end error message. The upgrade dialog embeds
the plans grid (shared with /plans), and the 402 upgrade_required path
now also covers enabling computer config, runtime deploy/switch, and
creating OpenClaw/Hermes/custom agents — which the API now gates before
creating or mutating anything, so free-plan attempts no longer leave
half-configured agents behind.
